Refactor initalization of Keycloak JS in Admin Console
