+++ js/apps/admin-ui/src/App.tsx
@@ -1,6 +1,4 @@
-import type KeycloakAdminClient from "@keycloak/keycloak-admin-client";
 import { Page } from "@patternfly/react-core";
-import type Keycloak from "keycloak-js";
 import { PropsWithChildren, Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import { Outlet } from "react-router-dom";
@@ -15,7 +13,7 @@ import { KeycloakSpinner } from "./components/keycloak-spinner/KeycloakSpinner";
 import { RealmsProvider } from "./context/RealmsContext";
 import { RecentRealmsProvider } from "./context/RecentRealms";
 import { AccessContextProvider } from "./context/access/Access";
-import { AdminClientContext } from "./context/auth/AdminClient";
+import { AdminClientProvider } from "./context/auth/AdminClient";
 import { RealmContextProvider } from "./context/realm-context/RealmContext";
 import { ServerInfoProvider } from "./context/server-info/ServerInfoProvider";
 import { WhoAmIContextProvider } from "./context/whoami/WhoAmI";
@@ -24,17 +22,8 @@ import { AuthWall } from "./root/AuthWall";
 
 export const mainPageContentId = "kc-main-content-page-container";
 
-export type AdminClientProps = {
-  keycloak: Keycloak;
-  adminClient: KeycloakAdminClient;
-};
-
-const AppContexts = ({
-  children,
-  keycloak,
-  adminClient,
-}: PropsWithChildren<AdminClientProps>) => (
-  <AdminClientContext.Provider value={{ keycloak, adminClient }}>
+const AppContexts = ({ children }: PropsWithChildren) => (
+  <AdminClientProvider>
     <WhoAmIContextProvider>
       <RealmsProvider>
         <RealmContextProvider>
@@ -50,12 +39,12 @@ const AppContexts = ({
         </RealmContextProvider>
       </RealmsProvider>
     </WhoAmIContextProvider>
-  </AdminClientContext.Provider>
+  </AdminClientProvider>
 );
 
-export const App = ({ keycloak, adminClient }: AdminClientProps) => {
+export const App = () => {
   return (
-    <AppContexts keycloak={keycloak} adminClient={adminClient}>
+    <AppContexts>
       <Page
         header={<Header />}
         isManagedSidebar
