+++ js/apps/admin-ui/src/context/auth/AdminClient.tsx
@@ -1,19 +1,24 @@
 import KeycloakAdminClient from "@keycloak/keycloak-admin-client";
-import Keycloak from "keycloak-js";
-import { DependencyList, useEffect } from "react";
+import { DependencyList, PropsWithChildren, useEffect } from "react";
 import { useErrorHandler } from "react-error-boundary";
-
-import environment from "../../environment";
 import { createNamedContext, useRequiredContext } from "ui-shared";
 
+import { adminClient } from "../../admin-client";
+
 export type AdminClientProps = {
-  keycloak: Keycloak;
   adminClient: KeycloakAdminClient;
 };
 
-export const AdminClientContext = createNamedContext<
-  AdminClientProps | undefined
->("AdminClientContext", undefined);
+const AdminClientContext = createNamedContext<AdminClientProps | undefined>(
+  "AdminClientContext",
+  undefined
+);
+
+export const AdminClientProvider = ({ children }: PropsWithChildren) => (
+  <AdminClientContext.Provider value={{ adminClient }}>
+    {children}
+  </AdminClientContext.Provider>
+);
 
 export const useAdminClient = () => useRequiredContext(AdminClientContext);
 
@@ -57,34 +62,4 @@ export function useFetch<T>(
 
     return () => controller.abort();
   }, deps);
-}
-
-export async function initAdminClient() {
-  const keycloak = new Keycloak({
-    url: environment.authServerUrl,
-    realm: environment.loginRealm,
-    clientId: environment.isRunningAsTheme
-      ? "security-admin-console"
-      : "security-admin-console-v2",
-  });
-
-  await keycloak.init({ onLoad: "check-sso", pkceMethod: "S256" });
-
-  const adminClient = new KeycloakAdminClient();
-
-  adminClient.setConfig({ realmName: environment.loginRealm });
-  adminClient.baseUrl = environment.authUrl;
-  adminClient.registerTokenProvider({
-    async getAccessToken() {
-      try {
-        await keycloak.updateToken(5);
-      } catch (error) {
-        keycloak.login();
-      }
-
-      return keycloak.token;
-    },
-  });
-
-  return { keycloak, adminClient };
 }
