+++ js/apps/admin-ui/src/routes.tsx
@@ -2,8 +2,6 @@ import type { AccessType } from "@keycloak/keycloak-admin-client/lib/defs/whoAmI
 import type { TFunction } from "i18next";
 import type { ComponentType } from "react";
 import type { NonIndexRouteObject, RouteObject } from "react-router";
-import { initAdminClient } from "./context/auth/AdminClient";
-import { initI18n } from "./i18n";
 
 import { App } from "./App";
 import { PageNotFoundSection } from "./PageNotFoundSection";
@@ -56,12 +54,8 @@ export const routes: AppRouteObject[] = [
   NotFoundRoute,
 ];
 
-const { keycloak, adminClient } = await initAdminClient();
-
-await initI18n(adminClient);
-
 export const RootRoute: RouteObject = {
   path: "/",
-  element: <App keycloak={keycloak} adminClient={adminClient} />,
+  element: <App />,
   children: routes,
 };
